Eliminate library loading overhead in colabfold_search
